added calendar view

+++ server/internal/transport/http/v1/goal_handler.go
@@ -4,6 +4,7 @@ import (
 	"log/slog"
 	"net/http"
 	"strconv"
+	"strings"
 
 	"github.com/ali-nur31/mile-do/internal/domain"
 	"github.com/ali-nur31/mile-do/internal/transport/http/v1/dto"
@@ -182,7 +183,7 @@ func (h *GoalHandler) UpdateGoal(c echo.Context) error {
 // @Param        id path int64 true "Goal ID"
 // @Success      201  {string}  map[string]string "goal has been removed"
 // @Failure      401  {object}  map[string]string "Unauthorized"
-// @Failure      404  {object}  map[string]string "Not Found"
+// @Failure      500  {object}  map[string]string "Internal Server Error"
 // @Failure      400  {object}  map[string]string "Bad Request"
 // @Router       /goals/{id} [delete]
 func (h *GoalHandler) DeleteGoalByID(c echo.Context) error {
@@ -196,10 +197,20 @@ func (h *GoalHandler) DeleteGoalByID(c echo.Context) error {
 		return c.JSON(http.StatusInternalServerError, map[string]string{"message": "internal server error", "error": err.Error()})
 	}
 
+	goal, err := h.service.GetGoalByID(c.Request().Context(), int64(id), int32(claims.ID))
+	if err != nil {
+		slog.Error("failed on getting goal by id for deletion", "error", err)
+		return c.JSON(http.StatusInternalServerError, map[string]string{"message": "internal server error", "error": err.Error()})
+	}
+
+	if strings.EqualFold(goal.Title, "routine") || strings.EqualFold(goal.Title, "other") {
+		return c.JSON(http.StatusBadRequest, map[string]string{"message": "bad request", "error": "cannot delete default tasks"})
+	}
+
 	err = h.service.DeleteGoalByID(c.Request().Context(), int64(id), int32(claims.ID))
 	if err != nil {
 		slog.Error("failed on deleting goal by id", "error", err)
-		return c.JSON(http.StatusNotFound, map[string]string{"message": "goal not found", "error": err.Error()})
+		return c.JSON(http.StatusInternalServerError, map[string]string{"message": "internal server error", "error": err.Error()})
 	}
 
 	return c.JSON(http.StatusOK, map[string]string{"message": "goal has been removed"})
